requirements.txt
absl-py==1.4.0
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
ansi2html==1.8.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
boto3==1.28.47
botocore==1.31.47
cachetools==5.3.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
chartjs==1.2
click==8.1.6
colorama==0.4.6
comm==0.1.3
common==0.1.2
contourpy==1.1.0
convertdate==2.4.0
cycler==0.11.0
Cython==0.29.34
dash==2.11.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.4
django-cors-headers==4.2.0
django-filter==23.2
django-storages==1.14
djangorestframework==3.14.0
emoji==2.7.0
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.16.3
Flask==2.2.5
flatbuffers==23.3.3
fonttools==4.41.1
fqdn==1.5.1
frozenlist==1.4.0
future==0.18.3
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.0
h5py==3.8.0
idna==3.4
ipykernel==6.23.0
ipython==8.13.2
ipython-genutils==0.2.0
ipywidgets==8.0.6
iso8601==1.1.0
isoduration==20.11.0
itsdangerous==2.1.2
jax==0.4.8
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
libopencv==0.0.1
Markdown==3.4.4
MarkupSafe==2.1.2
matplotlib==3.7.2
matplotlib-inline==0.1.6
mistune==2.0.5
ml-dtypes==0.1.0
multidict==6.0.4
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.2.0
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.72
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.0
plotly==5.15.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.22.4
psutil==5.9.5
psycopg2==2.9.7
psycopg2-binary==2.9.7
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pygame==2.4.0
Pygments==2.15.1
PyJWT==2.8.0
PyMeeus==0.5.12
pyparsing==3.0.9
pyrsistent==0.19.3
pyserial==3.5
pytesseract==0.3.10
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.3
QtPy==2.3.1
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
s3transfer==0.6.2
scipy==1.10.1
Send2Trash==1.8.2
serial==0.0.97
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.2
statsmodels==0.14.0
tenacity==8.2.2
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
tinycss2==1.2.1
tornado==6.3.1
traitlets==5.9.0
twilio==8.5.0
typing_extensions==4.7.1
tzdata==2023.3
uri-template==1.2.0
urllib3==2.0.4
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
widgetsnbextension==4.0.7
wrapt==1.14.1
yarl==1.9.2
